filtered the restaurant with normal javascript variable, it could not update in UI only updated in console.log
