+++ ReadMe.txt
@@ -1,1 +1,20 @@
 It is the React Web Learning
+
+# Food Ordering App Structure
+
+/*
+* Header
+    - Logo
+    - Nav Items
+* Body
+    - Search 
+    - RestrauntContainer
+        - RestrauntCard
+            - Img
+            - Name of Res,Cusine,Start Rating,Delivery Time
+* Footer
+    - CopyRight
+    - Links
+    - Address
+    - Contact
+*/
