+++ index.css
@@ -37,3 +37,6 @@
 .res-logo{
     width: 100%;
 }
+.filter-btn{
+    margin:10px;
+}
