+++ index.html
@@ -11,6 +11,6 @@
         <h1>Not Rendering</h1>
     </div>
     
-    <script type="module" src="App.js"></script>
+    <script type="module" src="./src/App.js"></script>
 </body>
 </html>
