+++ src/utils/constants.js
@@ -1,0 +1,2 @@
+export const LOGO_URL = "https://th.bing.com/th/id/OIP.LGetv97_2OwMpxHVYpn25gHaE8?w=260&h=180&c=7&r=0&o=5&pid=1.7";
+export const CDN_URL = "https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/" ;
